Reorganise package to move existing implementation to historic
