--- sbot/future/classless/__init__.py
@@ -1,3 +1,0 @@
-from .robot import arduino, comp, leds, motors, power, servos, utils, vision
-
-__all__ = ['arduino', 'comp', 'leds', 'motors', 'power', 'servos', 'utils', 'vision']
