+++ sbot/simulator/time_server.py
@@ -1,11 +1,13 @@
+"""A virtual clock source connected over a socket connection."""
+
 from __future__ import annotations
 
 import logging
 from datetime import datetime
 
-from sbot.exceptions import BoardDisconnectionError, IncorrectBoardError
-from sbot.serial_wrapper import SerialWrapper
-from sbot.utils import BoardIdentity, get_simulator_boards
+from ..internal.exceptions import BoardDisconnectionError, IncorrectBoardError
+from ..internal.serial_wrapper import SerialWrapper
+from ..internal.utils import BoardIdentity, get_simulator_boards
 
 logger = logging.getLogger(__name__)
 
@@ -13,6 +15,12 @@ BAUDRATE = 115200
 
 
 class TimeServer:
+    """
+    A virtual clock source connected over a socket connection.
+
+    :param serial_port: The URL of the serial port to connect to.
+    """
+
     @staticmethod
     def get_board_type() -> str:
         """
@@ -44,6 +52,11 @@ class TimeServer:
 
     @classmethod
     def initialise(cls) -> 'TimeServer' | None:
+        """
+        Initialise the board.
+
+        :return: The initialised board, or None if no board is found.
+        """
         # The filter here is the name of the emulated board in the simulator
         boards = get_simulator_boards('TimeServer')
 
@@ -84,10 +97,20 @@ class TimeServer:
         return BoardIdentity(*response.split(':'))
 
     def get_time(self) -> float:
+        """
+        Get the current time from the board.
+
+        :return: The current time in seconds since the epoch.
+        """
         time_str = self._serial.query('TIME?')
         return datetime.fromisoformat(time_str).timestamp()
 
     def sleep(self, duration: float) -> None:
+        """
+        Sleep for a specified duration.
+
+        :param duration: The duration to sleep for in seconds.
+        """
         if duration < 0:
             raise ValueError("sleep length must be non-negative")
